phot_kernels.py --> phot_kernels_in_situ.py and phot_kernels_merging.py --> phot_kernels.py
