+++ diffsky/experimental/kernels/__init__.py
@@ -3,6 +3,6 @@
 # flake8: noqa
 from .dbk_sed_kernels import _dbk_sed_kern
 from .dbk_photline_kernels import _mc_dbk_photline_kern_merging
-from .shambalam import _mc_phot_kern_merging
+from .phot_kernels import _mc_phot_kern_merging
 from .sed_kernels import _sed_kern
 from .photline_kernels import _mc_photline_kern_merging
