+++ diffsky/experimental/kernels/tests/test_phot_kernels_merging.py
@@ -7,7 +7,7 @@ from jax import random as jran
 
 from ....param_utils import diffsky_param_wrapper_merging as dpwm
 from ...tests import test_lightcone_generators as tlcg
-from .. import shambalam as gd_pkm
+from .. import phot_kernels as gd_pkm
 
 TOL = 1e-8
